Fetch data from the SISVAN API

fetch_demas_sisvan fetches data from DEMAS's (Departamento de Monitoramento, Avaliação e Disseminação de Informações Estratégicas em Saúde) SISVAN (Sistema de Vigilância Alimentar e Nutricional) API (Application Programming Interface).

fetch_demas_sisvan(
  codigo_municipio = NULL,
  uf = NULL,
  codigo_cnes = NULL,
  idade_minima = NULL,
  idade_maxima = NULL,
  codigo_fase_vida = NULL,
  codigo_povo_comunidade = NULL,
  codigo_escolaridade = NULL,
  ano_mes_competencia = NULL,
  gestante = NULL,
  limit = 20,
  offset = NULL
)

Arguments

codigo_municipio

(optional) an integer number with the city code (default: NULL).

uf

(optional) a string with the state code (default: NULL).

codigo_cnes

(optional) a string with the CNES (Cadastro Nacional de Estabelecimentos de Saúde) code (default: NULL).

idade_minima

(optional) an integer number with the minimum age (default: NULL).

idade_maxima

(optional) an integer number with the maximum age (default: NULL).

codigo_fase_vida

(optional) an integer number with the life phase code.

codigo_povo_comunidade

(optional) an integer number with the people/community code (default: NULL).

codigo_escolaridade

(optional) an integer number with the education code (default: NULL).

ano_mes_competencia

(optional) a string with the year and month of the competition (e.g., "202301") (default: NULL).

gestante

(optional) an integer number (0 or 1) representing the pregnancy status (0 for FALSE and 1 for TRUE) (default: NULL).

limit

(optional) an integer number with the maximum number of records to return (maximum value: 20) (default: 20).

offset

(optional) an integer number with the number of records to skip (default: NULL).

Value

A tibble object with the content of the API response.

Details

Learn more about the SISVAN API at:

https://opendatasus.saude.gov.br/dataset/sisvan-estado-nutricional

https://apidadosabertos.saude.gov.br/v1/

Examples

fetch_demas_sisvan(uf = "SP") |>
  dplyr::glimpse()
#> Rows: 20
#> Columns: 33
#> $ status_participacao                    <lgl> NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ codigo_municipio                       <int> 354780, 351380, 354530, 350100,…
#> $ uf                                     <chr> "SP", "SP", "SP", "SP", "SP", "…
#> $ municipio                              <chr> "SANTO ANDRE", "DIADEMA", "SALT…
#> $ codigo_cnes                            <chr> NA, "2801191", "2749254", "2094…
#> $ idade                                  <int> 31, 63, 57, 47, 67, 66, 1, 2, 6…
#> $ codigo_fase_vida                       <dbl> 7, 8, 7, 7, 8, 8, 2, 3, 4, 2, 1…
#> $ fase_vida                              <chr> "ADULTO", "IDOSO", "ADULTO", "A…
#> $ sexo                                   <chr> "F", "M", "F", "F", "F", "M", "…
#> $ codigo_raca_cor                        <chr> "02", "01", "01", "01", "99", "…
#> $ raca_cor                               <chr> "PRETA", "BRANCA", "BRANCA", "B…
#> $ codigo_povo_comunidade                 <lgl> NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ povo_comunidade                        <chr> "NÃO INFORMADO", "NÃO INFORMADO…
#> $ codigo_escolaridade                    <int> 99, NA, NA, NA, 99, NA, NA, NA,…
#> $ escolaridade                           <chr> "SEM INFORMAÇÃO", "SEM INFORMAÇ…
#> $ data_acompanhamento                    <chr> "2021-02-26", "2021-02-08", "20…
#> $ ano_mes_competencia                    <chr> "202102", "202102", "202102", "…
#> $ peso                                   <chr> "93,4", "94,6", "101,5", "90", …
#> $ altura                                 <chr> "160", "176", "151", "161", "15…
#> $ imc                                    <chr> "36,48", "30,54", "44,52", "34,…
#> $ imc_pre_gestacional                    <lgl> NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ peso_x_idade                           <chr> NA, NA, NA, NA, NA, NA, "Peso a…
#> $ peso_x_altura                          <chr> NA, NA, NA, NA, NA, NA, "Peso A…
#> $ crianca_altura_x_idade                 <chr> NA, NA, NA, NA, NA, NA, "Estatu…
#> $ crianca_imc_x_idade                    <chr> NA, NA, NA, NA, NA, NA, "Eutrof…
#> $ adolescente_altura_x_idade             <chr> NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ adolescente_imc_x_idade                <chr> NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ codigo_estado_nutricional_adulto       <chr> "Obesidade Grau II", NA, "Obesi…
#> $ codigo_estado_nutricional_idoso        <chr> NA, "Sobrepeso", NA, NA, "Adequ…
#> $ codigo_estado_nutricional_imc_gestante <chr> "Obesidade", NA, NA, NA, NA, NA…
#> $ codigo_sistema_origem_acompanhamento   <int> 2, 4, 4, 4, 4, 4, 4, 2, 2, 4, 4…
#> $ sistema_origem_acompanhamento          <chr> "AUXILIO BRASIL", "E-SUS AB", "…
#> $ codigo_sequencial_acompanhamento       <int> 1315131895, 1292981458, 1292981…