Fetch data from the SISVAN API

fetch_demas_sisvan fetches data from DEMAS's (Departamento de Monitoramento, Avaliação e Disseminação de Informações Estratégicas em Saúde) SISVAN (Sistema de Vigilância Alimentar e Nutricional) API (Application Programming Interface).

fetch_demas_sisvan(
  codigo_municipio = NULL,
  uf = NULL,
  codigo_cnes = NULL,
  idade_minima = NULL,
  idade_maxima = NULL,
  codigo_fase_vida = NULL,
  codigo_povo_comunidade = NULL,
  codigo_escolaridade = NULL,
  ano_mes_competencia = NULL,
  gestante = NULL,
  limit = 20,
  offset = NULL
)

Arguments

codigo_municipio

(optional) an integer number with the city code (default: NULL).

uf

(optional) a string with the state code (default: NULL).

codigo_cnes

(optional) a string with the CNES (Cadastro Nacional de Estabelecimentos de Saúde) code (default: NULL).

idade_minima

(optional) an integer number with the minimum age (default: NULL).

idade_maxima

(optional) an integer number with the maximum age (default: NULL).

codigo_fase_vida

(optional) an integer number with the life phase code.

codigo_povo_comunidade

(optional) an integer number with the people/community code (default: NULL).

codigo_escolaridade

(optional) an integer number with the education code (default: NULL).

ano_mes_competencia

(optional) a string with the year and month of the competition (e.g., "202301") (default: NULL).

gestante

(optional) an integer number (0 or 1) representing the pregnancy status (0 for FALSE and 1 for TRUE) (default: NULL).

limit

(optional) an integer number with the maximum number of records to return (maximum value: 20) (default: 20).

offset

(optional) an integer number with the number of records to skip (default: NULL).

Value

A tibble object with the content of the API response.

Details

Learn more about the SISVAN API at:

https://opendatasus.saude.gov.br/dataset/sisvan-estado-nutricional

https://apidadosabertos.saude.gov.br/v1/

Examples

fetch_demas_sisvan(uf = "SP") |>
  dplyr::glimpse()
#> Rows: 20
#> Columns: 33
#> $ status_participacao                    <lgl> NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ codigo_municipio                       <int> 354240, 351880, 351630, 354220,…
#> $ uf                                     <chr> "SP", "SP", "SP", "SP", "SP", "…
#> $ municipio                              <chr> "REGENTE FEIJO", "GUARULHOS", "…
#> $ codigo_cnes                            <chr> "2751151", "2062119", "2083809"…
#> $ idade                                  <int> 9, 31, 4, 3, 26, 57, 26, 65, 50…
#> $ codigo_fase_vida                       <dbl> 5, 7, 3, 3, 7, 7, 7, 8, 7, 8, 2…
#> $ fase_vida                              <chr> "ENTRE 7 ANOS A 10 ANOS", "ADUL…
#> $ sexo                                   <chr> "M", "F", "F", "M", "F", "F", "…
#> $ codigo_raca_cor                        <chr> "01", "04", "01", "01", "01", "…
#> $ raca_cor                               <chr> "BRANCA", "PARDA", "BRANCA", "B…
#> $ codigo_povo_comunidade                 <lgl> NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ povo_comunidade                        <chr> "NÃO INFORMADO", "NÃO INFORMADO…
#> $ codigo_escolaridade                    <int> 99, 99, 99, 99, 6, 5, 6, NA, NA…
#> $ escolaridade                           <chr> "SEM INFORMAÇÃO", "SEM INFORMAÇ…
#> $ data_acompanhamento                    <chr> "2016-02-02", "2016-02-29", "20…
#> $ ano_mes_competencia                    <chr> "201602", "201602", "201602", "…
#> $ peso                                   <chr> "51.00", "100", "15", "15.2", "…
#> $ altura                                 <chr> "174", "164", "75", "98", "1.59…
#> $ imc                                    <chr> "16.85", "37.18", "26.67", "15.…
#> $ imc_pre_gestacional                    <chr> NA, NA, NA, NA, NA, NA, "32.4",…
#> $ peso_x_idade                           <chr> "Peso elevado para a idade", NA…
#> $ peso_x_altura                          <chr> NA, NA, "Obesidade", "Peso Adeq…
#> $ crianca_altura_x_idade                 <chr> "Estatura adequada para a idade…
#> $ crianca_imc_x_idade                    <chr> "Eutrofia", NA, "Obesidade", "E…
#> $ adolescente_altura_x_idade             <chr> NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ adolescente_imc_x_idade                <chr> NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ codigo_estado_nutricional_adulto       <chr> NA, "Obesidade Grau II", NA, NA…
#> $ codigo_estado_nutricional_idoso        <chr> NA, NA, NA, NA, NA, NA, NA, "So…
#> $ codigo_estado_nutricional_imc_gestante <chr> NA, NA, NA, NA, NA, NA, "Adequa…
#> $ codigo_sistema_origem_acompanhamento   <int> 1, 2, 2, 2, 1, 1, 1, 4, 4, 4, 4…
#> $ sistema_origem_acompanhamento          <chr> "SISVAN WEB", "AUXILIO BRASIL",…
#> $ codigo_sequencial_acompanhamento       <int> 969128171, 978846659, 978880325…